+++ packages/DiffStudio/src/scripting-tools.ts
@@ -188,6 +188,7 @@ enum SCRIPT {
   SOLVER = `const solver = await grok.functions.eval('${PACKAGE_NAME}:${SOLVER_FUNC}');`,
   PREPARE = 'let call = solver.prepare({problem: odes, options: opts});',
   CALL = 'await call.call();',
+  FUNCTION_CALL = `let ${DF_NAME} = await grok.functions.call('${PACKAGE_NAME}:${SOLVER_FUNC}', {problem: odes, options: opts});`,
   OUTPUT = `let ${DF_NAME} = call.getParamValue('${DF_NAME}');`,
   SPACE2 = '  ',
   SPACE4 = '    ',
@@ -897,10 +898,11 @@ function getScriptMainBodyBasic(ivp: IVP): string[] {
   // 5. The 'call solver' lines
   res.push('');
   res.push(SCRIPT.SOLVER_COM);
-  res.push(SCRIPT.SOLVER);
-  res.push(SCRIPT.PREPARE);
-  res.push(SCRIPT.CALL);
-  res.push(SCRIPT.OUTPUT);
+  res.push(SCRIPT.FUNCTION_CALL);
+  // res.push(SCRIPT.SOLVER);
+  // res.push(SCRIPT.PREPARE);
+  // res.push(SCRIPT.CALL);
+  // res.push(SCRIPT.OUTPUT);
 
   return res;
 } // getScriptMainBodyBasic