+++ packages/Peptides/src/viewers/position-statistics-viewer.ts
@@ -5,6 +5,7 @@ import * as DG from 'datagrok-api/dg';
 import wu from 'wu';
 import $ from 'cash-dom';
 import {PeptideUtils} from '../peptideUtils';
+import {TAGS as bioTAGS} from '@datagrok-libraries/bio/src/utils/macromolecule';
 
 export const POSITION_HIDDEN_NAME = '~sequence_position_monomers';
 
@@ -29,6 +30,17 @@ export class SequencePositionStatsViewer extends DG.JsViewer {
     });
   }
 
+  getPositionFromColumn(): number {
+    const col = this.dataFrame.col(this.sequenceColumnName);
+    if (col == null)
+      return 0;
+    const positionString = col.getTag(bioTAGS.selectedPosition) ?? '1';
+    const position = parseInt(positionString);
+    if (Number.isNaN(position))
+      return 0;
+    return Math.max(0, position - 1);
+  }
+
   onTableAttached(): void {
     super.onTableAttached();
     if (this.dataFrame.columns.bySemType(DG.SEMTYPE.MACROMOLECULE) == null) {
@@ -44,7 +56,13 @@ export class SequencePositionStatsViewer extends DG.JsViewer {
 
     this.getProperty('sequenceColumnName')!.set(this, this.dataFrame.columns.bySemType(DG.SEMTYPE.MACROMOLECULE)!.name);
     this.getProperty('valueColumnName')!.set(this, wu(this.dataFrame.columns.numerical).next().value.name);
-    this.getProperty('position')!.set(this, 0);
+    this.getProperty('position')!.set(this, this.getPositionFromColumn());
+
+    this.subs.push(DG.debounce(this.dataFrame.onMetadataChanged, 200).subscribe((_) => {
+      const curPosition = this.getPositionFromColumn();
+      if (this.position !== curPosition)
+        this.getProperty('position')!.set(this, curPosition);
+    }));
   }
 
   render(): void {
@@ -67,11 +85,26 @@ export class SequencePositionStatsViewer extends DG.JsViewer {
 
     this._boxPlotViewer = this.dataFrame.plot.box({categoryColumnNames: [this._positionColumn.name], plotStyle: 'violin',
       valueColumnName: this.valueColumnName, colorColumnName: this._positionColumn.name, showColorSelector: false, showSizeSelector: false, showCategorySelector: false,
+      legendVisibility: DG.VisibilityMode.Never, description: `Position ${this.position + 1}`, markerColorColumnName: this._positionColumn.name, title: 'Sequence Position Statistics',
+      autoLayout: false, labelOrientation: 'Vert',
     });
+
+    setTimeout(() => {
+      this._boxPlotViewer!.props.title = 'Sequence Position Statistics';
+      this._boxPlotViewer!.props.description = `Position ${this.position + 1}`;
+    }, 200);
+
+    this._boxPlotViewer.props.statistics = ['min', 'max', 'avg', 'med', 'count'];
 
     this._boxPlotViewer.root.style.width = '100%';
     this._boxPlotViewer.root.style.height = '100%';
     this.root.appendChild(this._boxPlotViewer.root);
+    this._boxPlotViewer.sub(this._boxPlotViewer.onPropertyValueChanged.subscribe((_e) => {
+      if (this._boxPlotViewer?.props?.valueColumnName && this._boxPlotViewer?.props?.valueColumnName !== this.valueColumnName) {
+        const value = this._boxPlotViewer.props.valueColumnName;
+        setTimeout(() => this.getProperty('valueColumnName')!.set(this, value), 10);
+      }
+    }));
   }
 
   onPropertyChanged(property: DG.Property | null): void {