+++ tools/bin/utils/func-generation.ts
@@ -342,10 +342,19 @@ export function generateClassFunc(annotation: string, className: string, sep: st
   return annotation + `export function _${className}() {${sep}  return new ${className}();${sep}}${sep.repeat(2)}`;
 }
 
+const primitives = new Set([
+  'string',
+  'string[]',
+  'number',
+  'number[]',
+  'boolean',
+  'boolean[]',
+]);
+
 /** Generates a DG function. */
 export function generateFunc(annotation: string, funcName: string, sep: string = '\n', className: string = '', inputs: FuncParam[] = [], isAsync: boolean = false): string 
 {
-  let funcSigNature = (inputs.map((e)=>`${e.name}: ${e.type}`)).join(', ');
+  let funcSigNature = (inputs.map((e)=>`${e.name}: ${primitives.has(e.type ?? '') ? e.type : 'any' }`)).join(', ');
   let funcArguments = (inputs.map((e)=>e.name)).join(', ');
   
   return annotation + `export ${isAsync? 'async ': ''}function ${funcName}(${funcSigNature}) {${sep}  return ${className.length > 0 ? `${className}.` : ''}${funcName}(${funcArguments});${sep}}${sep.repeat(2)}`;