+++ packages/Bio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Bio changelog
 
+## 2.21.7 (2025-05-14)
+
+* Enable Header scrolling for non-MSA
+
+## 2.21.6 (2025-05-14)
+
+* Sequence position scrolling header
+
 ## 2.21.5 (2025-05-12)
 
 * Shifted sequence rendering support